Fully wrap tun task channel in strong type
